Commit of Project 3
Project 3 Overview

Project Objective: Build a six page website utilizing a single CSS
stylesheet to present a visual design that is consistent and effective
on all modern browsers and at least three device sizes. This site will
use the wireframes/web design/ux plan we created in a previous class.
Restrictions: HTML/CSS ONLY - Exception: Any javascript included with
framework is fine.
Time allotted: 5 days
NOTE: We had to use wireframes we created from another class and
basically turn them into a working website but ONLY on the design side.
So nothing is “functional” except for a few buttons to lead to certain
items created for the site. The site didn’t need to look exactly like
our wireframes but as close as we could with the time allotted. I
included the wireframes as well as the quick mini UX plan I did in the
previous class.